# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"facundo-villard","description":"facundo-villard","contact":{"name":"facundo_villard","email":"facundo.villard@mercadolibre.com"},"version":"last"},"servers":[{"url":"https://null_facundo-villard.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/sign-in":{"post":{"tags":["session-controller"],"operationId":"login","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountDTO"}}}}}}},"/api/v1/parts":{"post":{"tags":["part-controller"],"operationId":"updatePart","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStockDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockListResponseDTO"}}}}}}},"/api/v1/parts/orders":{"get":{"tags":["order-controller"],"operationId":"order_1","parameters":[{"name":"params","in":"query","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdersListResponseDTO"}}}}}},"post":{"tags":["order-controller"],"operationId":"order","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartsRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartsResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/orders/{orderNumberCM}":{"get":{"tags":["order-controller"],"operationId":"getOrderByNumber","parameters":[{"name":"orderNumberCM","in":"path","required":true,"schema":{"type":"string"}},{"name":"params","in":"query","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderStatusResponseDTO"}}}}}}},"/api/v1/parts/orders/{orderNumberCM}/total":{"get":{"tags":["order-controller"],"operationId":"getTotalOfOrderByNumber","parameters":[{"name":"orderNumberCM","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderTotalDTO"}}}}}}},"/api/v1/parts/list":{"get":{"tags":["part-controller"],"operationId":"getAllParts","parameters":[{"name":"params","in":"query","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartsListResponseDTO"}}}}}}}},"components":{"schemas":{"AccountDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"token":{"type":"string"}}},"UpdateStockDTO":{"required":["countryHouseId","partCode","quantity"],"type":"object","properties":{"countryHouseId":{"type":"integer","format":"int64"},"partCode":{"type":"string"},"quantity":{"minimum":0,"type":"integer","format":"int32"}}},"StockCountryHouseDTO":{"required":["partCode","quantity"],"type":"object","properties":{"partCode":{"pattern":"[0-9]{8}","type":"string"},"quantity":{"maximum":99999999,"type":"integer","format":"int32"}}},"StockListResponseDTO":{"type":"object","properties":{"stocks":{"type":"array","items":{"$ref":"#/components/schemas/StockCountryHouseDTO"}}}},"PartRequestDTO":{"type":"object","properties":{"partCode":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"urgent":{"type":"boolean"}}},"PartsRequestDTO":{"type":"object","properties":{"isUrgent":{"type":"boolean"},"parts":{"type":"array","items":{"$ref":"#/components/schemas/PartRequestDTO"}}}},"PartResponseDTO":{"type":"object","properties":{"partCode":{"type":"string"},"description":{"type":"string"},"maker":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"number","format":"float"},"total":{"type":"number","format":"float"}}},"PartsResponseDTO":{"type":"object","properties":{"total":{"type":"number","format":"double"},"parts":{"type":"array","items":{"$ref":"#/components/schemas/PartResponseDTO"}}}},"OrderDTO":{"type":"object","properties":{"orderNumber":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"deliveryDate":{"type":"string","format":"date"},"daysDelay":{"type":"integer","format":"int32"},"deliveryStatus":{"type":"string"},"orderDetails":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDTO"}}}},"OrderDetailDTO":{"type":"object","properties":{"partCode":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"accountType":{"type":"string"},"reason":{"type":"string"},"status":{"type":"string"}}},"OrdersListResponseDTO":{"type":"object","properties":{"dealerNumber":{"type":"integer","format":"int32"},"order":{"type":"array","items":{"$ref":"#/components/schemas/OrderDTO"}}}},"OrderStatusResponseDTO":{"type":"object","properties":{"orderNumberCE":{"type":"string"},"orderDate":{"type":"string","format":"date"},"orderStatus":{"type":"string"},"orderDetails":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDTO"}}}},"OrderTotalDTO":{"type":"object","properties":{"totalUrgentPrice":{"type":"number","format":"double"},"totalSalePrice":{"type":"number","format":"double"},"totalNormalPrice":{"type":"number","format":"double"}}},"PartDTO":{"required":["description","discountType","lastModification","longDimension","maker","netWeight","normalPrice","partCode","quantity","tallDimension","urgentPrice","widthDimension"],"type":"object","properties":{"partCode":{"pattern":"[0-9]{8}","type":"string"},"description":{"maxLength":100,"minLength":0,"pattern":"[a-zA-Z0-9]+","type":"string"},"maker":{"maxLength":100,"minLength":0,"pattern":"[a-zA-Z0-9]+","type":"string"},"quantity":{"maximum":99999999,"type":"integer","format":"int32"},"discountType":{"pattern":"[A-Z]{1}[0-9]{2}","type":"string"},"normalPrice":{"maximum":999999,"type":"number","format":"float"},"urgentPrice":{"maximum":999999,"type":"number","format":"float"},"netWeight":{"maximum":99999,"type":"number","format":"float"},"longDimension":{"maximum":9999,"type":"number","format":"float"},"widthDimension":{"maximum":9999,"type":"number","format":"float"},"tallDimension":{"maximum":9999,"type":"number","format":"float"},"lastModification":{"type":"string"}}},"PartsListResponseDTO":{"type":"object","properties":{"parts":{"type":"array","items":{"$ref":"#/components/schemas/PartDTO"}}}}}}}