# src/main/java/com/mercadolibre/facundo_villard/dtos/AccountDTO.java
package com.mercadolibre.facundo_villard.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class AccountDTO {
    private String username;
    private String password;
    private String token;
}